re-enable streaming in webui inference code
