+++ tools/webui/inference_engine/utils.py
@@ -1,4 +1,6 @@
+import io
 import html
+import wave
 from functools import partial
 from typing import Any, Callable
 
@@ -95,3 +97,16 @@ def build_html_error_message(error: Any) -> str:
         {html.escape(str(error))}
     </div>
     """
+
+
+def wav_chunk_header(sample_rate: int=44100, bit_depth: int=16, channels: int=1):
+    buffer = io.BytesIO()
+
+    with wave.open(buffer, "wb") as wav_file:
+        wav_file.setnchannels(channels)
+        wav_file.setsampwidth(bit_depth // 8)
+        wav_file.setframerate(sample_rate)
+
+    wav_header_bytes = buffer.getvalue()
+    buffer.close()
+    return wav_header_bytes
